rename obsolete filename
