+++ solidago/src/solidago/comparisons_to_scores/__init__.py
@@ -1,2 +1,2 @@
-from .indiv_matrix_inversion import HookeIndividualScores
+from .hooke_individual_scores import HookeIndividualScores
 from .continuous_bradley_terry import ContinuousBradleyTerry
